Adding a Grunt build system.
See the updated README for instructions on how to use the system.

This is based on the now closed PR #15 and addresses issue #8.

jsbeautifier is also setup to work but is not yet run on the repo.
Internally at CCP we still keep the project in a P4 repo and merging all
the formatting changes is not something we are yet prepared to do.