Added ccpwgl.getPendingLoads function
Returns a count of how many resources are still pending load. This function will help an end user to make a loading bar with percentage values during scene loads.

The .isLoading methods of ccpwgl space objects and the scene object return true when the base javascript object has loaded but not necessarily when all of it's resources have loaded (geometry, textures etc.).